change resources to custom resources
The following resources are now custom resources:
- `docker_installation`
- `docker_installation_package`
- `docker_installation_script`
- `docker_installation_tarball`
- `docker_service`
- `docker_service_base`
- `docker_service_manager`
- `docker_service_manager_execute`
- `docker_service_manager_systemd`

This fixes #1222 so the docker cookbook will work with Chef 18
